Autotools: detect libANN by the presence of ANN.h instead of using pkg-config
In theory libANN provides PC files and is findable/configurable via
`pkg-config`. In practice there is only one contemporary mainstream ecosystem
that packages libANN (Debian¹) and the package does not include PC files.² So a
more reasonable option is to detect libANN by the presence of its header. This
works out more successfully on both Debian-based operating systems as well as
systems that have a built-from-source-then-installed libANN.

Related to #1835.

¹ https://packages.debian.org/sid/libann-dev
² https://packages.debian.org/sid/amd64/libann-dev/filelist